+++ example/bin/alphabetic_order_example.dart
@@ -6,7 +6,7 @@ enum AlphabeticOrder with SerializeByName<AlphabeticOrder> {
 
   /// Reads a json map and returns the corresponding
   /// instance of `AlphabeticOrder`.
-  factory AlphabeticOrder.fromJson(Map<String, dynamic> json) =>
+  factory AlphabeticOrder.fromJson(Json json) =>
       SerializeByName.fromJson(json: json, values: values);
 }