+++ lib/src/mixin/serialize_by_index.dart
@@ -22,7 +22,7 @@ mixin SerializeByIndex<T extends Enum> on Enum implements Serializable {
   ///
   ///   /// Reads a json map and returns the corresponding
   ///   /// instance of `AlphabeticOrder`.
-  ///   factory AlphabeticOrder.fromJson(Map<String, dynamic> json) =>
+  ///   factory AlphabeticOrder.fromJson(Json json) =>
   ///   SerializeByIndex.fromJson(json: json, values: values);
   /// }
   /// ```
@@ -49,11 +49,11 @@ mixin SerializeByIndex<T extends Enum> on Enum implements Serializable {
   ///   const key = 'myCustomKey';
   ///
   ///   @override
-  ///   Map<String, dynamic> toJson() => {key: index};
+  ///   Map<String, Object?> toJson() => {key: index};
   ///
   ///   /// Reads a json map and returns the corresponding
   ///   /// instance of `AlphabeticOrder`.
-  ///   factory AlphabeticOrder.fromJson(Map<String, dynamic> json) =>
+  ///   factory AlphabeticOrder.fromJson(Json json) =>
   ///   SerializeByIndex.fromJsonCustomKey(
   ///     json: json,
   ///     values: values,
@@ -66,17 +66,15 @@ mixin SerializeByIndex<T extends Enum> on Enum implements Serializable {
     required List<E> values,
     required String key,
   }) {
-    if (json.containsKey(key) &&
-        json[key] is int &&
-        json[key] < values.length &&
-        json[key] >= 0) {
-      return values[json[key]];
-    } else {
-      throw ExceptionOf<E>(
-          message: 'Json validation error',
-          expectedState: '{$key: a valid index '
-              'between 0 and ${values.length - 1}}',
-          invalidState: '$json');
+    switch (json[key]) {
+      case int index when index >= 0 && index < values.length:
+        return values[index];
+      default:
+        throw ExceptionOf<E>(
+            message: 'Json validation error',
+            expectedState: '{$key: a valid index '
+                'between 0 and ${values.length - 1}}',
+            invalidState: '$json');
     }
   }
 }