+++ test/serialize_by_index_custom_key_test.dart
@@ -1,4 +1,5 @@
 // ignore_for_file: unused_local_variable
+
 import 'package:serialize_enum/serialize_enum.dart';
 import 'package:test/test.dart';
 
@@ -9,11 +10,11 @@ enum AlphabeticOrder implements SerializeByIndex {
   static const key = 'customKey';
 
   @override
-  Map<String, dynamic> toJson() => {key: index};
+  Json toJson() => {key: index};
 
   /// Reads a json map and returns the corresponding
   /// instance of `AlphabeticOrder`.
-  factory AlphabeticOrder.fromJson(Map<String, dynamic> json) =>
+  factory AlphabeticOrder.fromJson(Json json) =>
       SerializeByIndex.fromJsonCustomKey(json: json, values: values, key: key);
 }