+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+## 0.1.2
+- Changed the typedef `Json` from `Map<String, dynamic>`
+  to `Map<String, Object?>`.
+- Updated dependencies.
+
 ## 0.1.1
 - Added interface `Serializable`.
 - Enums with `SerializeByName` or `SerializeByIndex` now
-  implement `Serializable`. 
+  implement `Serializable`.
 
 ## 0.1.0
 - Initial version.